Introduce batch size to improve cnn performance
Introducing a batch size.
In the paper, you claim a training advantage of Quantum NN over classical NN.
In section 4.2 Fig 3b, we see that the classical neural network has a much larger loss than the quantum network.

However, this is merely because you don't use a batch_size (i.e., you use the whole data in one batch).
If you were to introduce a batch size, the classical net drastically outperforms all the others. Loss reaches ~0.

The core of the issue is that you have a training issue in your classical net. (The reason is that your classical net is too deep for such a simple problem, which is why you run into this problem.)